improve&refactor: abstract the common chat module and improve scrollToBottom fn usage
