+++ README.md
@@ -7,7 +7,7 @@
 
 ### 项目展示：
 
-github对gif图有限制，我就直接截图了，具体详情建议直接[线上体验](https://im.aermin.top)。给个star呀~
+github对gif图有限制，我就直接截图了，具体详情建议直接[线上体验](https://im.aermin.top)。***可以的话给个star支持一下*** 😄
 
 ![image](https://user-images.githubusercontent.com/24861316/53351929-e1d33300-395c-11e9-84a9-0a9fd793b5a1.png)
 
@@ -153,6 +153,8 @@ ps: 本地发表情和发文件和github登录无法使用，需要自己去gith
   - [x] 发表情
   - [x] 发文件
   - [x] 下载文件
+  - [ ] 图片方法查看
+  - [ ] @某人/Enter快捷键发送信息
   - [ ] 支持Markdown
   - [ ] 支持Quote
